+++ CHANGELOG.md
@@ -5,7 +5,10 @@
 - self-heating correction
 - time shift detection fft
 
-* Refactored `HampelDaytimeNighttime` outlier removal method, it now runs 100x faster (23)
+* Refactored `HampelDaytimeNighttime` outlier removal method, it now runs 100x faster. Also added parameter
+  `use_differencing` to calculate outliers from the double-differenced time series instead of the original data (23).
+  Also added parameter `separate_day_night` to run the filter without the separation into daytime/nighttime data. The
+  original Hampel class is therefore now implemented here and was removed (24).
 
 ## v0.90.0 | XX Nov 2025
 
@@ -89,6 +92,7 @@
 * (21) `diive.pkgs.createvar.air.aerodynamic_resistance`
 * (22) `diive.pkgs.fluxprocessingchain.level31_storagecorrection.FluxStorageCorrectionSinglePointEddyPro`
 * (23) `diive.pkgs.outlierdetection.hampel.HampelDaytimeNighttime`
+* (24) `diive.pkgs.outlierdetection.hampel.Hampel`
 
 ## v0.89.0 | 23 Jul 2025